ic_udpifc: index the cursor IC history table with uint32
addCursorIcEntry() computes its bucket as a uint32, but getCursorIcEntry(),
updateCursorIcEntry(), pruneCursorIcEntry() and purgeCursorIcEntry() used a
uint8 index. The table size comes from gp_interconnect_cursor_ic_table_size,
which defaults to 128 but may be raised up to 102400, and the GUC's own
description tells users to raise it when UDFs with many concurrent cursors
hang. Any value above 255 makes lookups and updates miss their bucket, so
mismatched packets are no longer acknowledged, and turns the prune and purge
loops into infinite loops while holding ic_control_info.lock, which also
wedges the receiver thread. Use uint32 throughout.

+++ src/backend/cdb/motion/ic_udpifc.c
@@ -989,7 +989,7 @@ static void
 updateCursorIcEntry(CursorICHistoryTable *t, uint32 icId, uint8 status)
 {
 	struct CursorICHistoryEntry *p;
-	uint8		index = icId % t->size;
+	uint32		index = icId % t->size;
 
 	for (p = t->table[index]; p; p = p->next)
 	{
@@ -1010,7 +1010,7 @@ static CursorICHistoryEntry *
 getCursorIcEntry(CursorICHistoryTable *t, uint32 icId)
 {
 	struct CursorICHistoryEntry *p;
-	uint8		index = icId % t->size;
+	uint32		index = icId % t->size;
 
 	for (p = t->table[index]; p; p = p->next)
 	{
@@ -1030,7 +1030,7 @@ getCursorIcEntry(CursorICHistoryTable *t, uint32 icId)
 static void
 pruneCursorIcEntry(CursorICHistoryTable *t, uint32 icId)
 {
-	uint8		index;
+	uint32		index;
 
 	for (index = 0; index < t->size; index++)
 	{
@@ -1079,7 +1079,7 @@ pruneCursorIcEntry(CursorICHistoryTable *t, uint32 icId)
 static void
 purgeCursorIcEntry(CursorICHistoryTable *t)
 {
-	uint8		index;
+	uint32		index;
 
 	for (index = 0; index < t->size; index++)
 	{
